Change module default param for port_fetch_method
Previously the 'poudriere ports' command would fetch the ports using
subversion, which created a requirement on subversion being available on
the system.

This change uses the default value from the manpage of 'poudriere' to
fetch the ports using 'portsnap'.  This should mean users are able to
take advantage of the defaults with fewer parameter changes, or
dependency package installations.