Prefer optional non-empty strings
Instead of using empty strings for optional values, change the data type
to Optional[String[1]] to ensure that either a valid String is provided
or no value is passwd.

A side effect is that these parameters cannot be set explicitely undef
by the end user if they have a default value, and $poudriere::zpool is
concerned by this.  But since this default value is bogus, change it to
undef to avoid the issue.

This is therefore a breaking change.