Rework $poudriere::env::arch
poudriere(8) defaults to the architecure of the host, so it makes sense
to do so in the module.  On amd64, it is possible to specify another
architecture for cross-building packages, but it is an exception, not
the rule, so existing users should be unaffected by this change.

While here, explicitely list allowed values.