Sync poudriere.conf template with recent release
Legacy parameters where removed in a previous commit, here we focus on
updating the template to match the default configuration file.  We
ensure that each comment line is replicated verbatim, mandatory
parameters values are set to the default value by the poudriere class
and optional values have a dedicated conditional block under the setting
description.