Fix portstree detection
With this change, we use the same method to check if a portstree is
present or absent: query poudriere for the list of port trees and search
into that list.

The previous way of looking for a ports tree presence by testing the
existence of a directory does not work when using poudriere::portstree's
mountpoint parameter:

```puppet
class profile::poudriere (
  String $zpool,
) {
  class { 'poudriere':
    zpool           => $zpool,
    freebsd_host    => 'http://ftp.FreeBSD.org/',
    ccache_enable   => true,
    use_portlint    => 'yes',
    allow_make_jobs => 'yes',
  }

  class { 'poudriere::xbuild':
  }

  poudriere::portstree { 'default':
    ensure       => present,
    fetch_method => 'null',
    mountpoint   => '/home/romain/FreeBSD/ports',
  }
}
```

While here, tune the poudriere ports parameters to only list portstree
names without metadata not column titles.